Revert "Rafli"

+++ app/Http/Controllers/CartController.php
@@ -85,8 +85,6 @@ class CartController extends Controller
     {
         // return $request;
         cart::create($request->all());
-        notify()->success('item berhasil ditambahkan');
-
         return redirect()->back();
     }
 
@@ -117,16 +115,19 @@ class CartController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($id)
+    public function destroy(Cart $cart,$p)
     {
         // mencari data user & member
         $user = auth()->user()->id;
         $member = member::where('user_id', $user)->pluck('id')->first();
 
-
         // mencari produk di tabel cart menggunakkan id member lalu dihapus
-        $item = cart::where('member_id', $member)->where('produk_id', $id)->first();
-        $item->delete();
+        $item = cart::where('member_id', $member)->where('id', $cart->id)->first();
+        if ($item) {
+            $item->delete();
+        }else{
+            return $p;
+        }
         return redirect()->back();
     }
